Move watch utils into separate module
